requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
Django==4.1.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.20.1
sqlparse==0.4.3
tomli==2.0.1
django-betterforms==2.0.0
